# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This service returns information about lifejourneys",
        "title": "Benefit Service",
        "contact": {},
        "version": "1.0"
    },
    "host": "[TBD]",
    "basePath": "/",
    "paths": {
        "/healthcheck": {
            "get": {
                "description": "Returns Healthy",
                "summary": "Returns Healthy",
                "operationId": "healthcheck",
                "responses": {
                    "200": {
                        "description": "Healthy",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/lifejourney": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Request Info on Life Journey",
                "operationId": "Life Journey",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/renderings.LifeJourneyResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/renderings.QPSError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/renderings.QPSError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/renderings.QPSError"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.LifeJourney": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "description_fr": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "related_benefits": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "title": {
                    "type": "string"
                },
                "title_fr": {
                    "type": "string"
                }
            }
        },
        "renderings.LifeJourneyResponse": {
            "type": "object",
            "properties": {
                "lifejourneys": {
                    "description": "Life Journey ID",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.LifeJourney"
                    }
                }
            }
        },
        "renderings.QPSError": {
            "type": "object",
            "properties": {
                "error_code": {
                    "type": "string"
                }
            }
        }
    }
}